Flatted directory structure, rename files
