+++ indra_network_service/templates/query_template.html
