+++ indra_network_service/tests/test_api.py
@@ -1,7 +1,7 @@
 import logging
 import unittest
 
-from service_api.api import app
+from indra_network_service.api import app
 
 
 logger = logging.getLogger('api test')
